requirements.txt
# Core dependencies
pyrofork==2.6.1
tgcrypto==1.2.5
python-dotenv==1.0.1
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
frozenlist==1.4.0
multidict==6.0.4
yarl==1.9.3
pytz==2024.2
GitPython==3.1.43
psutil==5.9.8
ujson==5.9.0
uvloop==0.19.0